change file position
