+++ lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 // import 'package:reading_app/style/theme.dart';
-import 'package:reading_app/pages/bookshelf/bookshelf.dart';
 import 'pages/all.dart';
 import 'theme/theme.dart';
 import 'ui/icons.dart';
